# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BCIS Subscription Billing and Collection System API",
    "description": "Official REST API for the BCIS Subscription Billing and Collection System",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {}
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "System"
        ],
        "description": "Liveness probe verifying that the API server is responsive",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string"
                    },
                    "uptime": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "status",
                    "timestamp",
                    "uptime"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/ready": {
      "get": {
        "tags": [
          "System"
        ],
        "description": "Readiness probe verifying that database and critical subsystems are healthy",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "database": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "database",
                    "timestamp"
                  ]
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "database": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "database",
                    "timestamp"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/health": {
      "get": {
        "tags": [
          "System"
        ],
        "description": "Liveness probe verifying that the API server is responsive",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string"
                    },
                    "uptime": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "status",
                    "timestamp",
                    "uptime"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ready": {
      "get": {
        "tags": [
          "System"
        ],
        "description": "Readiness probe verifying that database and critical subsystems are healthy",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "database": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "database",
                    "timestamp"
                  ]
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "database": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "database",
                    "timestamp"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Authenticate user and issue session Bearer token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "minLength": 1
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "username": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string",
                          "nullable": true
                        }
                      },
                      "required": [
                        "id",
                        "username",
                        "displayName"
                      ]
                    },
                    "token": {
                      "type": "string"
                    },
                    "expiresAt": {
                      "type": "string"
                    },
                    "roles": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "permissions": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "user",
                    "token",
                    "expiresAt",
                    "roles",
                    "permissions"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Revoke active session token",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "Retrieve currently authenticated user identity and effective permissions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "username": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string",
                          "nullable": true
                        }
                      },
                      "required": [
                        "id",
                        "username",
                        "displayName"
                      ]
                    },
                    "roles": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "permissions": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "user",
                    "roles",
                    "permissions"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/audit-check": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "Admin verification endpoint requiring user.manage permission",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/subscribers": {
      "get": {
        "tags": [
          "Subscribers"
        ],
        "description": "List subscribers with search, status filtering, and pagination",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "search",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "default": 1
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "default": 20
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "tags": [
          "Subscribers"
        ],
        "description": "Create a new subscriber with primary address and authoritative account number",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "firstName",
                  "lastName",
                  "primaryContactNumber",
                  "address"
                ],
                "properties": {
                  "firstName": {
                    "type": "string",
                    "minLength": 1
                  },
                  "middleName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string",
                    "minLength": 1
                  },
                  "businessName": {
                    "type": "string"
                  },
                  "primaryContactNumber": {
                    "type": "string",
                    "minLength": 7
                  },
                  "secondaryContactNumber": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "notes": {
                    "type": "string"
                  },
                  "address": {
                    "type": "object",
                    "required": [
                      "line1",
                      "barangay"
                    ],
                    "properties": {
                      "label": {
                        "type": "string",
                        "default": "Home"
                      },
                      "line1": {
                        "type": "string",
                        "minLength": 1
                      },
                      "line2": {
                        "type": "string"
                      },
                      "barangay": {
                        "type": "string",
                        "minLength": 1
                      },
                      "cityMunicipality": {
                        "type": "string",
                        "default": "Malaybalay City"
                      },
                      "province": {
                        "type": "string",
                        "default": "Bukidnon"
                      },
                      "postalCode": {
                        "type": "string",
                        "default": "8700"
                      },
                      "landmark": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/subscribers/{id}": {
      "get": {
        "tags": [
          "Subscribers"
        ],
        "description": "Get full subscriber profile, addresses, service accounts, and history",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "patch": {
        "tags": [
          "Subscribers"
        ],
        "description": "Update subscriber details and primary address",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/subscribers/{id}/service-accounts": {
      "post": {
        "tags": [
          "Subscribers"
        ],
        "description": "Create a billable service account under a subscriber with catalog rate snapshot",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "servicePlanId"
                ],
                "properties": {
                  "servicePlanId": {
                    "type": "string"
                  },
                  "installationAddressId": {
                    "type": "string"
                  },
                  "activationDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "billingStartDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "billingDay": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 28
                  },
                  "dueDay": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 28
                  },
                  "collectorId": {
                    "type": "string"
                  },
                  "collectionAreaId": {
                    "type": "string"
                  },
                  "reason": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/service-accounts/{id}/status": {
      "patch": {
        "tags": [
          "Subscribers"
        ],
        "description": "Change service account status (suspend, reconnect, disconnect) with mandatory audit reason",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "toStatus",
                  "reason"
                ],
                "properties": {
                  "toStatus": {
                    "type": "string",
                    "enum": [
                      "PENDING",
                      "ACTIVE",
                      "SUSPENDED",
                      "DISCONNECTED",
                      "TERMINATED"
                    ]
                  },
                  "reason": {
                    "type": "string",
                    "minLength": 3
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/service-plans": {
      "get": {
        "tags": [
          "Subscribers"
        ],
        "description": "List active catalog service plans",
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/collection-areas": {
      "get": {
        "tags": [
          "Subscribers"
        ],
        "description": "List active collection areas",
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/collectors": {
      "get": {
        "tags": [
          "Subscribers"
        ],
        "description": "List active collectors",
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
  "tags": [
    {
      "name": "System",
      "description": "Health, readiness, and system status"
    },
    {
      "name": "Auth",
      "description": "Authentication and session management"
    },
    {
      "name": "Subscribers",
      "description": "Subscriber profile and account management"
    },
    {
      "name": "Billing",
      "description": "Invoices, billing cycles, and ledger"
    },
    {
      "name": "Payments",
      "description": "Payment processing and allocation"
    },
    {
      "name": "Collections",
      "description": "Collector batches and remittances"
    },
    {
      "name": "Receivables",
      "description": "AR aging and overdue tracking"
    },
    {
      "name": "Reports",
      "description": "Financial and operational reports"
    }
  ]
}